% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-package.R
\docType{package}
\name{traj-package}
\alias{traj}
\alias{traj-package}
\title{traj: Clustering of Functional Data Based on Measures of Change}
\description{
Implements a three-step procedure in the spirit of Leffondree et al. (2004) to identify clusters of individual longitudinal trajectories. The procedure involves (1) computing a number of "measures of change" capturing various features of the trajectories; (2) using a Principal Component Analysis based dimension reduction algorithm to select a subset of measures and (3) using the k-means clustering algorithm to identify clusters of trajectories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=traj}
}

}
\author{
\strong{Maintainer}: Laurence Boulanger \email{laurence.boulanger@umontreal.ca}

Authors:
\itemize{
  \item Marie-Pierre Sylvestre \email{marie-pierre.sylvestre@umontreal.ca}
}

Other contributors:
\itemize{
  \item Gillis Delmas Tchouangue Dinkou [contributor]
  \item Dan Vatnik [contributor]
}

}
\keyword{internal}
